chore: rename package powerbi_ext → powerbi_extension; bump python to >=3.10; add LICENSE
